+++ plugins/graphql-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @frontside/backstage-plugin-graphql-backend
 
+## 0.1.1
+
+### Patch Changes
+
+- 33222ff: Update @frontside/hydraphql to 0.1.1
+- Updated dependencies [33222ff]
+  - @frontside/backstage-plugin-graphql-backend-node@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes